+++ tests/testthat/doctestExample-tester.R
@@ -1,0 +1,2 @@
+
+some_code_that_might_go_in_an_example()