+++ R/docs.R
@@ -237,6 +237,8 @@ NULL
 #'
 #' At present, `@doctestExample` doesn't add any code to the tests.
 #'
+#' `@doctestExample` was added after doctest 0.2.0.
+#'
 #' @name doctestExample-tag
 #' @aliases @doctestExample
 NULL