+++ NEWS.md
@@ -1,3 +1,7 @@
+# doctest (development version)
+
+* Added a `@doctestExample` tag as a drop-in replacement for `@example`.
+
 # doctest 0.2.0
 
 * First CRAN release.